move watchdogs into subfolder
