+++ browser_use/browser/watchdogs/aboutblank_watchdog.py
