+++ browser_use/browser/watchdogs/crash_watchdog.py
