+++ browser_use/browser/watchdogs/default_action_watchdog.py
