+++ browser_use/browser/watchdogs/permissions_watchdog.py
