+++ browser_use/browser/watchdogs/popups_watchdog.py
