+++ browser_use/browser/watchdogs/security_watchdog.py
